Release v2.6.9 - Eighth-round deep analysis fixes (ISS-2025-0192)
13 fixes across 13 files for unification correctness, precision, and robustness:
- ListTerm.unify() rollback on partial failure
- Union/3 Set1 deduplication
- Clause/2 TermCopier variable renaming
- SumList/MaxList/MinList integer precision and first-element init
- Between/3 long precision
- PutCode/1 supplementary Unicode
- Tab/1 negative validation
- TermCopier PrologString handling
- ListTerm resolveBindings optimization
- Read/1 EOF safety

Tests: 429/429 JUnit, 20/20 examples pass.

+++ src/main/java/it/denzosoft/jprolog/builtin/arithmetic/Between.java
@@ -71,11 +71,13 @@ public class Between implements BuiltIn {
         Term resolvedValueTerm = valueTerm.resolveBindings(bindings);
 
         if (resolvedValueTerm instanceof Variable) {
+            // START_CHANGE: ISS-2025-0192 - Use long constructor to avoid double precision loss
             // Generate all values in the range
             boolean foundSolution = false;
             for (long i = low; i <= high; i++) {
                 Map<String, Term> newBindings = new HashMap<>(bindings);
-                if (valueTerm.unify(new it.denzosoft.jprolog.core.terms.Number((double) i), newBindings)) {
+                if (valueTerm.unify(new it.denzosoft.jprolog.core.terms.Number(i), newBindings)) {
+            // END_CHANGE: ISS-2025-0192
                     solutions.add(newBindings);
                     foundSolution = true;
                 }
